+++ db/migrate/20211125193821_add_image_url_to_wines.rb
@@ -1,0 +1,5 @@
+class AddImageUrlToWines < ActiveRecord::Migration[6.0]
+  def change
+    add_column :wines, :image_url, :string
+  end
+end
