+++ db/migrate/20211125200345_add_rating_to_wines.rb
@@ -1,0 +1,5 @@
+class AddRatingToWines < ActiveRecord::Migration[6.0]
+  def change
+    add_column :wines, :rating, :integer
+  end
+end
