+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_11_24_213351) do
+ActiveRecord::Schema.define(version: 2021_11_25_200345) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -56,6 +56,11 @@ ActiveRecord::Schema.define(version: 2021_11_24_213351) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "saq_id"
+    t.float "latitude"
+    t.float "longitude"
+    t.string "origin"
+    t.string "image_url"
+    t.integer "rating"
   end
 
   add_foreign_key "food_pairings", "foods"
